requirements.txt
python-telegram-bot>=21.0
openai>=1.12.0
python-dotenv>=1.0.0
aiohttp>=3.11.0
aiosqlite>=0.19.0
yookassa>=2.3.4